requirement.txt
Package               Version
--------------------- -------
asgiref               3.7.2
crispy-bootstrap5     0.7
Django                4.2.4
django-crispy-forms   2.0
django-jazzmin        2.6.0
django-recaptcha      3.0.0
django-shortuuidfield 0.1.3
django-uuidfield      0.5.0
Pillow                10.0.0
pip                   22.3.1
setuptools            65.5.0
shortuuid             1.0.11
six                   1.16.0
sqlparse              0.4.4
tzdata                2023.3